fix up to be ready for redo
