+++ browser/react/index.js
